Add a timeout to waiting for the garbage collector to cleanup files.
git-svn-id: http://svn.cern.ch/guest/lhcb/LHCb/trunk@130606 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: aa3bb8cfd8bacd69a4eabb2eeeac8bc377c1c11f

+++ LHCb/Tools/FileStager/src/gc/garbage.cpp
@@ -20,7 +20,8 @@ std::string tmpdir;
 void cleanup( int )
 {
    fs::path p( tmpdir );
-   fs::remove_all( tmpdir );
+   if ( fs::exists(p) )
+      fs::remove_all( tmpdir );
    exit( 0 );
 }
 #endif
@@ -68,6 +69,10 @@ int main( int argc, char* argv[] )
    } catch ( const boost::bad_any_cast& e ) {
       cout << "dir is not a correct string" << endl;
       cout << "usage: garbage.exe pid dir" << endl;
+      return -2;
+   }
+   if ( !fs::exists( fs::path( tmpdir ) ) ) {
+      cout << "Directory " << tmpdir << " does not exist!" << endl;
       return -2;
    }
